FirstApplication renaming + misc. improvements
This commit gives the application a more suitable name that reflects
its purpose: from FirstApplication to CountWords.

It also adds CLI parameters validation, logging and a README.md file
explaining how to run the application and output examples.

Finally, it automates the output directory cleanup to remove the need
of manually deleting such directory to avoid RuntimeExceptions is it
was not done.