Renamed sample readme file & updated README.md to reflect this
This commits renames the README.md sample file provided in the app to
avoid github reading and displaying it automatically which is confusing
as its content has nothing to do the app. It's only provided to be used
as input for the app (if desired).

The commit also updates the actual app's README.md file to reflect
this change in the example it includes.